initial service restructure
