+++ SPC-2021/angular/src/app/services/api services/client.service.ts
@@ -1,0 +1,20 @@
+import { Injectable } from '@angular/core';
+import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
+import { Observable, of } from 'rxjs';
+import { environment } from 'src/environments/environment';
+
+const endpoint = environment.apiUrl;
+
+const httpOptions = {
+  headers: new HttpHeaders({
+    'Content-Type':  'application/json'
+  })
+};
+
+@Injectable({
+  providedIn: 'root'
+})
+export class ClientService 
+{
+  constructor(private http: HttpClient) { }
+}
