+++ SPC-2021/angular/src/app/services/gathering services/mission-board-gatherer.service.ts
@@ -2,12 +2,12 @@ import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { BoardEntry, EntryType } from 'src/classes/missionBoards';
 import { MissionBoardAPIService } from './santa-api.service';
-import { MissionMapper } from './mapper.service';
+import { MissionMapper } from '../utility services/mapper.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class MissionBoardService {
+export class MissionBoardGathererService {
 
 constructor(private missionBoardAPIService: MissionBoardAPIService, private missionMapper: MissionMapper) { }
