+++ SPC-2021/angular/src/app/services/utility services/countries.service.ts
@@ -1,6 +1,6 @@
   
 import { Injectable } from '@angular/core';
-import * as countries from "../../assets/data/countries.json"
+import * as countries from "../../../assets/data/countries.json"
 
 @Injectable({
   providedIn: 'root'
