+++ SPC-2021/angular/src/app/services/utility services/exporter.service.ts
